server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Yoshio Miyamae <yoshio.miyamae@famoce-succellion.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "<4"
django-cors-headers = "^3.10.1"
graphene-django = "^2.15.0"
django-graphql-jwt = "^0.3.4"
graphene = "<3"
Pillow = "^9.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
mypy = "^0.931"
flake8 = "^4.0.1"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
